pyproject.toml
[project]
name = "autocode"
version = "0.1.0"
description = "Edge-native AI coding assistant — local-first, deterministic-first"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "textual>=0.89",
    "ollama>=0.6.1",
    "prompt_toolkit>=3.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
]

[project.scripts]
autocode = "autocode.cli:app"
hybridcoder = "autocode.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-pyyaml>=6.0",
    "httpx>=0.27",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
]
layer1 = ["tree-sitter>=0.25.2", "tree-sitter-python>=0.25.0"]
layer2 = ["lancedb>=0.29", "sentence-transformers>=5.0"]
layer3 = ["llama-cpp-python>=0.3", "outlines>=0.1"]
layer4 = ["ollama>=0.4"]
all = ["autocode[layer1,layer2,layer3,layer4]"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autocode"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["autocode"]
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services (deselected by default)",
    "benchmark: marks benchmark tests",
]
addopts = ""

[dependency-groups]
dev = [
    "types-pyyaml>=6.0",
]
